requirements.txt
numpy
pillow
pycuda=2021.1